Refactor messy qso list code into qsoListComponent (#16)
Part of #1 

Following up on #15 convert qso list code into a separate component.

This leaves only handling of new qso submission in `main` - it requires interaction between `qsoFormComponent` and `qsoListComponent`. I do not yet have a solution to this problem, so leaving it there, but it should be solved soon.